requirements.txt
Django>=4.2
djangorestframework
python-dotenv
faster-whisper==0.7.0
openai==0.28
drf-yasg
